added edit plan route and complete it's ui functionality, and corrected the issue about deleting plans

+++ api/controller/plan.controller.js
@@ -22,6 +22,7 @@ export const createPlan = async (req, res, next) => {
     const newPlan = new Plans({
       ...req.body,
       incomeId: income._id,
+      incomeSlug: income.slug,
       currency: income.currency,
     });
 
@@ -34,11 +35,24 @@ export const createPlan = async (req, res, next) => {
 
 export const getPlans = async (req, res, next) => {
   try {
-    const incomeId = req.params.incomeId;
-    const plans = await Plans.find({ incomeId: incomeId });
-    if (!plans) {
-      return next(errorHandler(404, "Plan not found"));
-    }
+    // const incomeId = req.params.incomeId;
+
+    const startIndex = parseInt(req.query.startIndex) || 0;
+    const limit = parseInt(req.query.limit) || 10;
+    const sortDirection = req.query.order === "asc" ? 1 : -1;
+    // if (!plans) {
+    //   return next(errorHandler(404, "Plan not found"));
+    // }
+    const plans = await Plans.find({
+      ...(req.query.planId && { _id: req.query.planId }),
+      ...(req.query.incomeId && { incomeId: req.query.incomeId }),
+      ...(req.query.name && {
+        $or: [{ name: { $regex: req.query.name, $options: "i" } }],
+      }),
+    })
+      .sort({ updatedAt: sortDirection })
+      .skip(startIndex)
+      .limit(limit);
     res.status(200).json(plans);
   } catch (error) {
     next(error);
@@ -48,12 +62,43 @@ export const getPlans = async (req, res, next) => {
 export const deletePlan = async (req, res, next) => {
   try {
     const plan = await Plans.findById(req.params.planId);
+
+    if (req.user.id !== req.params.userId) {
+      return next(errorHandler(403, "You are not allowed to delete this Plan"));
+    }
     if (!plan) {
       return next(errorHandler(403, "Plan does not exist"));
     }
 
     await Plans.findByIdAndDelete(req.params.planId);
     res.status(200).json("Deleted successfully");
+  } catch (error) {
+    next(error);
+  }
+};
+
+export const editPlan = async (req, res, next) => {
+  try {
+    const plan = await Plans.findById(req.params.planId);
+    if (req.user.id !== req.params.userId) {
+      return next(errorHandler(403, "You are not allowed to delete this Plan"));
+    }
+    if (!plan) {
+      next(errorHandler(404, "Plan does not exist"));
+    }
+
+    const edited = await Plans.findByIdAndUpdate(
+      req.params.planId,
+      {
+        $set: {
+          name: req.body.name,
+          description: req.body.description,
+          amount: req.body.amount,
+        },
+      },
+      { new: true }
+    );
+    res.status(200).json(edited);
   } catch (error) {
     next(error);
   }
